requirements.txt
celery
django
djangorestframework
gdal
numpy
rasterio
requests
scipy